Correct StripFrom Right
This commit implements a change to detect with StripFrom Right
exists in the parameters and return the fragment.
It also implements a change where StripFrom is in the parameters
but neither Left or Right is, and returns a failure.

To test the change, deploy the macro and launch the string_example.yaml
template.